Renamed classes from Normalizer (specific) to Processors (generic)
