+++ markdown_chunkify/__init__.py
@@ -1,5 +1,5 @@
-from markdown_chunkify.components.normalizers import GeminiNormalizer
 from markdown_chunkify.components.parsers import PyMuPDFMParser
+from markdown_chunkify.components.processors import GeminiNormalizer
 from markdown_chunkify.components.splitters import MarkdownSplitter
 
 __all__ = ["MarkdownSplitter", "PyMuPDFMParser", "GeminiNormalizer"]
